+++ Sources/Sentry/SentryWatchdogTerminationScopeObserver.m
@@ -79,7 +79,7 @@
 
 - (void)setTags:(nullable NSDictionary<NSString *, NSString *> *)tags
 {
-    // Left blank on purpose
+    [self.attributesProcessor setTags:tags];
 }
 
 - (void)setUser:(nullable SentryUser *)user
@@ -89,7 +89,9 @@
 
 - (void)setTraceContext:(nullable NSDictionary<NSString *, id> *)traceContext
 {
-    // Left blank on purpose
+    // Nothing to do here, Trace Context is not persisted for watchdog termination events
+    // On regular events, we have the current trace in memory, but there isn't time to persist one
+    // in watchdog termination events
 }
 
 @end